#54 - Arranged slices by entity/method
